+++ source/pcrgeo/geo_rasterspacetest.cc
@@ -56,7 +56,7 @@ BOOST_AUTO_TEST_CASE(nr_cells)
 {
   using namespace geo;
 
-  BOOST_CHECK(d_rs1->nrCells() == 11 * 12);
+  BOOST_CHECK(d_rs1->nrCells() == static_cast<size_t>(11 * 12));
 }
